Bug Fixes and Massive UI Overhaul
(massive commit due to git corruption)

Code has been changed to properly support mergcap and the propper
merging and cutting of pcap files. Pcap files are condensed by the
program to only contain the important and relevant information needed
to grab the handshake (beacon frames are not erased). SIGINT handler
implemented to handle most FILE* and other memory allocation that was
initialized at the start of runtime but may not have been freed when
interupt occured such as AP data and wireless device data.

Major UI overhaul using ncurses allows a more streamlined UI. Colored
and columned texted is easily written and erased to not flood the
terminal buffer and not leave excess amounts of text on screen from
runtime. Overall look improvements as well.

Data structure refactoring. BSSID has been refactored to be a union
so that the bssid can be read as well as individual portions such as
OUI and NIC. While this has not been put to practice in current
iterations of code, I believe it best practice to make it easier to
solve for future problems. AP and Client stuctures have been re-worked
to better store their own relevant data to mirror new attack throttling